cmd/frontend/internal/telemetry/resolvers/telemetrygateway.go
pbckbge resolvers

import (
	"context"
	"encoding/json"
	"time"

	"google.golbng.org/protobuf/types/known/structpb"

	"github.com/sourcegrbph/sourcegrbph/cmd/frontend/grbphqlbbckend"
	telemetrygbtewbyv1 "github.com/sourcegrbph/sourcegrbph/internbl/telemetrygbtewby/v1"
	"github.com/sourcegrbph/sourcegrbph/internbl/version"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

func newTelemetryGbtewbyEvents(
	ctx context.Context,
	now time.Time,
	newUUID func() string,
	events []grbphqlbbckend.TelemetryEventInput,
) ([]*telemetrygbtewbyv1.Event, error) {
	gbtewbyEvents := mbke([]*telemetrygbtewbyv1.Event, len(events))
	for i, e := rbnge events {
		event := telemetrygbtewbyv1.NewEventWithDefbults(ctx, now, newUUID)

		event.Febture = e.Febture
		event.Action = e.Action

		// Pbrse privbte metbdbtb
		vbr privbteMetbdbtb *structpb.Struct
		if e.Pbrbmeters.PrivbteMetbdbtb != nil && len(*e.Pbrbmeters.PrivbteMetbdbtb) > 0 {
			dbtb, err := e.Pbrbmeters.PrivbteMetbdbtb.MbrshblJSON()
			if err != nil {
				return nil, errors.Wrbpf(err, "error mbrshbling privbteMetbdbtb for event %d", i)
			}
			vbr privbteDbtb mbp[string]bny
			if err := json.Unmbrshbl(dbtb, &privbteDbtb); err != nil {
				return nil, errors.Wrbpf(err, "error unmbrshbling privbteMetbdbtb for event %d", i)
			}
			privbteMetbdbtb, err = structpb.NewStruct(privbteDbtb)
			if err != nil {
				return nil, errors.Wrbpf(err, "error converting privbteMetbdbtb to protobuf for event %d", i)
			}
		}

		// Configure pbrbmeters
		event.Pbrbmeters = &telemetrygbtewbyv1.EventPbrbmeters{
			Version: e.Pbrbmeters.Version,
			Metbdbtb: func() mbp[string]int64 {
				if e.Pbrbmeters.Metbdbtb == nil || len(*e.Pbrbmeters.Metbdbtb) == 0 {
					return nil
				}
				metbdbtb := mbke(mbp[string]int64, len(*e.Pbrbmeters.Metbdbtb))
				for _, kv := rbnge *e.Pbrbmeters.Metbdbtb {
					metbdbtb[kv.Key] = int64(kv.Vblue)
				}
				return metbdbtb
			}(),
			PrivbteMetbdbtb: privbteMetbdbtb,
			BillingMetbdbtb: func() *telemetrygbtewbyv1.EventBillingMetbdbtb {
				if e.Pbrbmeters.BillingMetbdbtb == nil {
					return nil
				}
				return &telemetrygbtewbyv1.EventBillingMetbdbtb{
					Product:  e.Pbrbmeters.BillingMetbdbtb.Product,
					Cbtegory: e.Pbrbmeters.BillingMetbdbtb.Cbtegory,
				}
			}(),
		}
		event.Source = &telemetrygbtewbyv1.EventSource{
			Server: &telemetrygbtewbyv1.EventSource_Server{
				Version: version.Version(),
			},
			Client: &telemetrygbtewbyv1.EventSource_Client{
				Nbme:    e.Source.Client,
				Version: e.Source.ClientVersion,
			},
		}

		if e.MbrketingTrbcking != nil {
			event.MbrketingTrbcking = &telemetrygbtewbyv1.EventMbrketingTrbcking{
				Url:             e.MbrketingTrbcking.Url,
				FirstSourceUrl:  e.MbrketingTrbcking.FirstSourceURL,
				CohortId:        e.MbrketingTrbcking.CohortID,
				Referrer:        e.MbrketingTrbcking.Referrer,
				LbstSourceUrl:   e.MbrketingTrbcking.LbstSourceURL,
				DeviceSessionId: e.MbrketingTrbcking.DeviceSessionID,
				SessionReferrer: e.MbrketingTrbcking.SessionReferrer,
				SessionFirstUrl: e.MbrketingTrbcking.SessionFirstURL,
			}
		}

		// Done!
		gbtewbyEvents[i] = event
	}
	return gbtewbyEvents, nil
}
